# utilities/Swagutils/SampleWebApi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SampleWebApi",
    "description": "This is a sample web api for testing / demo purposes for Swagabond.",
    "termsOfService": "https://en.wikipedia.org/wiki/Lorem_ipsum",
    "contact": {
      "name": "Hugh Man",
      "url": "https://www.google.com",
      "email": "test@gmail.com",
      "x-contactPhoneNumber": "(123)-456-7890"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT",
      "x-licenseDescription": "The MIT License allows anyone to use, modify, and distribute the software for any purpose, as long as the original license and copyright notice are included."
    },
    "version": "v1",
    "x-infoTwitterHandle": "@madeUpTwitterHandle"
  },
  "paths": {
...
    "/api/v1/menuitems": {
      "get": {
        "tags": [
          "MenuItem"
        ],
        "summary": "Get a big ole list of every menu item in the whole system.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.MenuItemResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.MenuItemResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.MenuItemResponse"
                }
              }
            }
          }
        },
        "x-operationExtension": "hello world"
      },
      "post": {
        "tags": [
          "MenuItem"
        ],
        "summary": "Add a new menu item",
        "requestBody": {
          "description": "The menu item request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleWebApi.Controllers.MenutItemPostRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleWebApi.Controllers.MenutItemPostRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SampleWebApi.Controllers.MenutItemPostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.MenuItemResponseItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.MenuItemResponseItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.MenuItemResponseItem"
                }
              }
            }
          }
        },
        "x-operationExtension": "hello world"
      }
    },
    "/api/v1/menuitems/{id}": {
      "get": {
        "tags": [
          "MenuItem"
        ],
        "summary": "Get a single menu item by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "the menu item's ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "x-schemaExtension": "hello world"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.MenuItemResponseItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.MenuItemResponseItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.MenuItemResponseItem"
                }
              }
            }
          }
        },
        "x-operationExtension": "hello world"
      }
    },
    "/api/v1/menuitems/{id}/full": {
      "get": {
        "tags": [
          "MenuItem"
        ],
        "summary": "Same as GetById, but includes franchise information as well.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The menu item ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "x-schemaExtension": "hello world"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.FullMenuItemGetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.FullMenuItemGetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.FullMenuItemGetResponse"
                }
              }
            }
          }
        },
        "x-operationExtension": "hello world"
      }
    },
    "/api/v1/restaurants": {
      "get": {
        "tags": [
          "Restaurant"
        ],
        "summary": "Get a list of every single restaurant in the system",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantGetResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantGetResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantGetResponse"
                }
              }
            }
          }
        },
        "x-operationExtension": "hello world"
      },
      "post": {
        "tags": [
          "Restaurant"
        ],
        "summary": "Create a new restaurant",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantPostRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantPostRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantPostRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantGetResponseItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantGetResponseItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantGetResponseItem"
                }
              }
            }
          }
        },
        "x-operationExtension": "hello world"
      }
    },
    "/api/v1/restaurants/{id}": {
      "get": {
        "tags": [
          "Restaurant"
        ],
        "summary": "Get a single restaurant by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "x-schemaExtension": "hello world"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantGetResponseItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantGetResponseItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantGetResponseItem"
                }
              }
            }
          }
        },
        "x-operationExtension": "hello world"
      },
      "delete": {
        "tags": [
          "Restaurant"
        ],
        "summary": "Deletes a single restaurant by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "x-schemaExtension": "hello world"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "x-operationExtension": "hello world"
      }
    }
  },
  "components": {
    "schemas": {
...
      "SampleWebApi.Controllers.FranchiseInformation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique id of the franchise",
            "format": "uuid",
            "x-schemaExtension": "hello world"
          },
          "name": {
            "type": "string",
            "description": "The name of the franchise",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "slogan": {
            "type": "string",
            "description": "The slogan of the franchise, in english.",
            "nullable": true,
            "x-schemaExtension": "hello world"
          }
        },
        "additionalProperties": false,
        "description": "An object containing franchise information",
        "x-schemaExtension": "hello world"
      },
      "SampleWebApi.Controllers.FullMenuItemGetResponse": {
        "type": "object",
        "properties": {
          "item": {
            "$ref": "#/components/schemas/SampleWebApi.Controllers.MenuItemResponseItem"
          },
          "franchise": {
            "$ref": "#/components/schemas/SampleWebApi.Controllers.FranchiseInformation"
          }
        },
        "additionalProperties": false,
        "description": "A full menu item response, as well as the franchise info",
        "x-schemaExtension": "hello world"
      },
      "SampleWebApi.Controllers.MenuItemResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SampleWebApi.Controllers.MenuItemResponseItem"
            },
            "description": "The list of menu items",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "totalCount": {
            "type": "integer",
            "description": "The total count of menu items",
            "format": "int32",
            "x-schemaExtension": "hello world"
          }
        },
        "additionalProperties": false,
        "description": "A response containing multiple menu items",
        "x-schemaExtension": "hello world"
      },
      "SampleWebApi.Controllers.MenuItemResponseItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The menu item's ID",
            "format": "uuid",
            "x-schemaExtension": "hello world"
          },
          "franchiseId": {
            "type": "string",
            "description": "The franchises unique ID",
            "format": "uuid",
            "x-schemaExtension": "hello world"
          },
          "name": {
            "type": "string",
            "description": "The menu item name",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "description": {
            "type": "string",
            "description": "The menu item description (english)",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "nutritionFacts": {
            "$ref": "#/components/schemas/SampleWebApi.Controllers.MenutItemNutritionFacts"
          }
        },
        "additionalProperties": false,
        "description": "A single menu item",
        "x-schemaExtension": "hello world"
      },
      "SampleWebApi.Controllers.MenutItemNutritionFacts": {
        "type": "object",
        "properties": {
          "calories": {
            "type": "integer",
            "description": "How many calories the menu item contains",
            "format": "int32",
            "x-schemaExtension": "hello world"
          },
          "proteinNutritionFacts": {
            "$ref": "#/components/schemas/SampleWebApi.Controllers.ProteinNutritionFact"
          }
        },
        "additionalProperties": false,
        "description": "A set of menu nutrition facts",
        "x-schemaExtension": "hello world"
      },
      "SampleWebApi.Controllers.MenutItemPostRequest": {
        "type": "object",
        "properties": {
          "franchiseId": {
            "type": "string",
            "description": "The franchise ID this menu item is offered by",
            "format": "uuid",
            "x-schemaExtension": "hello world"
          },
          "name": {
            "type": "string",
            "description": "The name of the menu item",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "description": {
            "type": "string",
            "description": "The menu item's description in english",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "calories": {
            "type": "integer",
            "description": "How many calories is this menu item?",
            "format": "int32",
            "x-schemaExtension": "hello world"
          },
          "proteinGrams": {
            "type": "integer",
            "description": "How many grams of protein does the item contain?",
            "format": "int32",
            "x-schemaExtension": "hello world"
          }
        },
        "additionalProperties": false,
        "description": "A post request for menu items",
        "x-schemaExtension": "hello world"
      },
      "SampleWebApi.Controllers.ProteinNutritionFact": {
        "type": "object",
        "properties": {
          "proteinGrams": {
            "type": "integer",
            "description": "Grams of protein",
            "format": "int32",
            "x-schemaExtension": "hello world"
          },
          "proteinMilligrams": {
            "type": "integer",
            "description": "Milligrams of protein",
            "format": "int32",
            "x-schemaExtension": "hello world"
          }
        },
        "additionalProperties": false,
        "description": "Specific nutrition facts related to protein",
        "x-schemaExtension": "hello world"
      },
      "SampleWebApi.Controllers.RestaurantGetResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SampleWebApi.Controllers.RestaurantGetResponseItem"
            },
            "description": "List of restaurants",
            "nullable": true,
            "x-schemaExtension": "hello world"
          }
        },
        "additionalProperties": false,
        "description": "The response containing the list of restuarants",
        "x-schemaExtension": "hello world"
      },
      "SampleWebApi.Controllers.RestaurantGetResponseItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier for the restaurant.",
            "format": "uuid",
            "x-schemaExtension": "hello world"
          },
          "franchiseId": {
            "type": "string",
            "description": "The Franchise ID for the restaurant.",
            "format": "uuid",
            "x-schemaExtension": "hello world"
          },
          "storeNumber": {
            "type": "integer",
            "description": "Unique number franchises use to identify the restaurant.",
            "format": "int32",
            "x-schemaExtension": "hello world"
          },
          "address": {
            "type": "string",
            "description": "The physical address of the restaurant",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "zip": {
            "type": "string",
            "description": "The 5 digit postal code of the restaurant.",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "city": {
            "type": "string",
            "description": "The city the restaurant is located in.",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "state": {
            "$ref": "#/components/schemas/SampleWebApi.Entities.State"
          }
        },
        "additionalProperties": false,
        "description": "A single restaurant item, which includes the restaurant's name, address, phone number, website, and description.",
        "x-schemaExtension": "hello world"
      },
      "SampleWebApi.Controllers.RestaurantPostRequest": {
        "type": "object",
        "properties": {
          "franchiseId": {
            "type": "string",
            "description": "The Franchise ID for the restaurant.",
            "format": "uuid",
            "x-schemaExtension": "hello world"
          },
          "storeNumber": {
            "type": "integer",
            "description": "Unique number franchises use to identify the restaurant.",
            "format": "int32",
            "x-schemaExtension": "hello world"
          },
          "address": {
            "type": "string",
            "description": "The physical address of the restaurant",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "zip": {
            "type": "string",
            "description": "The 5 digit postal code of the restaurant.",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "city": {
            "type": "string",
            "description": "The city the restaurant is located in.",
            "nullable": true,
            "x-schemaExtension": "hello world"
          },
          "state": {
            "$ref": "#/components/schemas/SampleWebApi.Entities.State"
          }
        },
        "additionalProperties": false,
        "description": "Request for creating a new restaurant",
        "x-schemaExtension": "hello world"
      },
      "SampleWebApi.Entities.State": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12,
          13,
          14,
          15,
          16,
          17,
          18,
          19,
          20,
          21,
          22,
          23,
          24,
          25,
          26,
          27,
          28,
          29,
          30,
          31,
          32,
          33,
          34,
          35,
          36,
          37,
          38,
          39,
          40,
          41,
          42,
          43,
          44,
          45,
          46,
          47,
          48,
          49,
          50,
          101,
          102,
          103,
          104,
          105,
          106,
          107
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Alabama",
          "Alaska",
          "Arizona",
          "Arkansas",
          "California",
          "Colorado",
          "Connecticut",
          "Delaware",
          "Florida",
          "Georgia",
          "Hawaii",
          "Idaho",
          "Illinois",
          "Indiana",
          "Iowa",
          "Kansas",
          "Kentucky",
          "Louisiana",
          "Maine",
          "Maryland",
          "Massachusetts",
          "Michigan",
          "Minnesota",
          "Mississippi",
          "Missouri",
          "Montana",
          "Nebraska",
          "Nevada",
          "NewHampshire",
          "NewJersey",
          "NewMexico",
          "NewYork",
          "NorthCarolina",
          "NorthDakota",
          "Ohio",
          "Oklahoma",
          "Oregon",
          "Pennsylvania",
          "RhodeIsland",
          "SouthCarolina",
          "SouthDakota",
          "Tennessee",
          "Texas",
          "Utah",
          "Vermont",
          "Virginia",
          "Washington",
          "WestVirginia",
          "Wisconsin",
          "Wyoming",
          "AmericanSamoa",
          "Guam",
          "NorthernMarianaIslands",
          "PuertoRico",
          "USMinorOutlyingIslands",
          "USVirginIslands",
          "DistrictOfColumbia"
        ],
        "x-schemaExtension": "hello world"
      }
    }
  }
}